Parts 1-3 done included in p1 and ptest updated
ptest was updated to include test for calculating/verifying SSE (we were given it in the slides, but wanted to confirm I was doing it right).  Although that is isn't the ASE, would be simple to change that, just take SSE/n, where n equals number of examples, in the case of the slides, n=15.

p1 has been updated to find the ASE of training and test data with the dummy variable (part 1 and 2, after of course applying w to the matrix, which is multiple each value of w to a column of X).  It has also been updated to do part 3, which is the same as before except without a dummy variable.  Working on doing part 4 of linear regression problem.