Problem 1 almost complete.
Problem 1 now does everything it is supposed to do (part1-3 already done previously), and now creates a graph of the ASE's from d=2..10 and also a nice d=0 (where there is no dummy variable found earlier, as an extra aside).  p1part4.png is this graph.  Note: Each time it is run it will be slightly different.