Starting/base file for part 2 Decision Tree
Starting/base file for part 2, problem 1.  It takes in the training csv data and has all of the contents in two variables (X, Y), where the first column of knn_train.csv is in Y (it is all the 1s and -1s, the class label) and X has all the features, like we did in the last assignment.  We can change it to something else if anyone wants, but thought I'd do it similar to last assignment for this step.  

It does NOT do problem 1 of part 2, yet.