Decision stump working...
I believe it finds a correct decision stump.  (Doesn't even come close to answering number 1 to part 2, need to add code that will keep track of the labels of the two branches, it's there/being kept track of, but need to output it).

Also need to normalize the data (which I wasn't doing).