p2 Might be working for part 1&2
Part 1 and 2 might be complete...?  It now makes 10 images and the mean image.  
If this is correct, then all we need to do is figure out how to do part 3.  

Part 3 is a little confusing as I don't understand: "Identify for each dimension the image that has the largest value in that dimension and plot it using imshow"