Got it sorta working (for one subject)
It seems to being doing something for a single subject, I need to change it though to work with all the training subject; I haven't ran this on the general pop tests yet, as I suspect it will be ridiculously low (one subject, and not much with figuring out what decision tree values to tweak, such as depth, number of leaves, number of features, etc.)  

I also need to play with which columns might need to be removed.  I am pretty sure the time (column 0) will need to be removed, because I believe it's possible it might be leading to over fitting. (We'll see about it).