It runs...
This "runs", which is for the group population set.  It doesn't guess ever from the guesses that I see, it is relatively close to guessing 1 on some of them, but it never does.  I haven't done anything to calculate the "true" accuracy (binary prediction: precision, recall, or F1 measure, nor probabilistic predictions: area under ROC curve).