Now using CORRECT probabilities
Instead of area under ROC for probabilities it is actually supposed to be some p1 and p2 values we come up with.  In my case the random decision tree from sklearn has a way to output what these probabilities are, which I am not using.

NOTE: This is not ready to work with the test cases, because they are of different format!  instead of 7 columns across it is now 63 columns, all so each instance will fit on ONE line, which this program doesn't account for (yet).