Individual subject 2 results and code
The code for subject 2 individual tests for decision tree and its output which is basically always guessing 0.  This happened because it didn't have enough data to work with and was forced to not use "warm start" which made it do very poorly (terrible).  But it is what we have for this model.